+++ lib/commands/addClients.d.ts
@@ -1,2 +1,0 @@
-import { CommandEvent } from "../types";
-export declare const addClients: (commandEvent: CommandEvent) => Promise<any>;