+++ changelog.md
@@ -19,9 +19,12 @@
 
 ### Changed
 
+- [adjust_inventory.ts] handle 0 stock
+
 ### Fixed
 
 - fix bug, get data from repzo with query per_page = 50000 @maramalshen
+- [adjust_inventory.ts] fix bug: get measureunit from the variant.product.sv_measureUnit @maramalshen
 
 ### Removed